Add base TestFixture classes for all edit/play mode tests
These classes have methods tagged with [SetUp] and [TearDown] for setup/teardown logic that is identical across all tests: currently, the calling of (Edit|Play)ModeTestHelpers.ResetScene. This means we never have to remember to explicitly call that method in tests again! We just have to remember to inherit the base TestFixture classes...