Add an IAppEnvironment service interface
This abstraction allows clients to learn what named environment they are running in, without duplicating environment names across those clients in the Inspector.

Only current implementation is a ScriptableObject, as a game's environment is not scene-specific. Also updated RemoteConfigConfigurationSource to inject this new service, rather than accepting an environment name in the Inspector.