PlayerPrefsLocalPreferences no longer adds app version to pref keys
Most prefs are not actually going to change between versions, and concatenating the version completely breaks users' expectations on Android when they update the APK and get a message that their settings will not be affected. If consumers ever need a version-specific pref then they can concatenate the version themselves.