Add an ApplicationQuitter script
I would've sworn we already had one...

This script is kinda unique in that I want it to exit play mode in the Editor, but exit the application in built players. Thus, the simplest implementation is with `#if UNITY_EDITOR`, but that's only meaningful if compiled in Unity. So this is the first script in UnityUtil that's _not_ part of UnityUtil.dll!