Split out Serilog-specific classes from `Unity.Extensions.Logging`
...into new `Unity.Extensions.Serilog` package.

This allows UnityUtil packages to use the `ILoggerFactory.CreateLogger(UnityEngine.Object)` extension method from `Unity.Extensions.Logging` without a dependency on Serilog types. It also allows package consumers to work with the Serilog extension methods without using MEL, if so inclined.

Also replace `UnityLogContext` usage with `ValueTuple<UnityEngine.Object>` so that the struct doesn't have to be moved to a common project for both `UnityObjectLogger` and `UnityLogContextDestructuringPolicy` to reference, while still not enforcing a new destructuring policy for all `UnityEngine.Object` properties